+++ core/src/main/java/be/orbinson/aem/opentelemetry/core/services/AttributeLogRecordProcessor.java
@@ -4,13 +4,13 @@ import io.opentelemetry.api.common.AttributeKey;
 import io.opentelemetry.context.Context;
 import io.opentelemetry.sdk.logs.LogRecordProcessor;
 import io.opentelemetry.sdk.logs.ReadWriteLogRecord;
-import org.apache.commons.lang3.StringUtils;
 
 public class AttributeLogRecordProcessor implements LogRecordProcessor {
 
     @Override
     public void onEmit(Context context, ReadWriteLogRecord readWriteLogRecord) {
-        String hostname = StringUtils.defaultIfBlank(System.getenv("HOSTNAME"), "localhost");
+        String env = System.getenv("HOSTNAME");
+        String hostname = env != null && !env.trim().isEmpty() ? env : "localhost";
         readWriteLogRecord.setAttribute(AttributeKey.stringKey("host"), hostname);
         readWriteLogRecord.setAttribute(
                 AttributeKey.stringKey("logger"),