+++ core/src/test/java/be/orbinson/aem/opentelemetry/core/ContextPropagatorTest.java
@@ -1,8 +1,8 @@
 package be.orbinson.aem.opentelemetry.core;
 
 import be.orbinson.aem.opentelemetry.core.ContextPropagator;
-import io.wcm.testing.mock.aem.junit5.AemContext;
-import io.wcm.testing.mock.aem.junit5.AemContextExtension;
+import org.apache.sling.testing.mock.sling.junit5.SlingContext;
+import org.apache.sling.testing.mock.sling.junit5.SlingContextExtension;
 import org.apache.sling.testing.mock.sling.servlet.MockSlingHttpServletRequest;
 import org.apache.sling.testing.mock.sling.servlet.MockSlingHttpServletResponse;
 import org.junit.jupiter.api.Test;
@@ -14,11 +14,11 @@ import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
 
-@ExtendWith(AemContextExtension.class)
+@ExtendWith(SlingContextExtension.class)
 class ContextPropagatorTest {
     public static final String TRACEPARENT_HEADER = "traceparent";
     public static final String TRACEPARENT_VALUE = "00-80e1afed08e019fc1110464cfa66635c-7a085853722dc6d2-01";
-    private final AemContext context = new AemContext();
+    private final SlingContext context = new SlingContext();
 
     @Test
     void traceParentIsSet() {